requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
paho-mqtt>=2.0.0
opencv-python>=4.9.0
face_recognition>=1.3.0
oracledb>=2.2.0
openpyxl>=3.1.2
numpy>=1.26.0
Pillow>=10.3.0
python-multipart>=0.0.9
